Verify intuition about temperature sampling
The literature suggests modifying the softmax output layer in order to
do temperature sampling, but all of the implementations I found online
do some extra math on the *output* of the softmax layer, rather than its
input logits.

This verifies that the transformation of the output is equivalent to the
literature's temperature transformation of the input logit values.