Use high DPI PNGs and add Bierce corpus
Generated SVGs are _far_ too large for the webpage.
Want to compare the haiku corpus with a regular one.
This changed the conclusion from Zipf's law holds, to Zipf's law kind of
holds.
Also, the pinned version of the en_core_web_sm language model doesn't
work with the latest SpaCy, and the bug that forced downloading the
model from GitHub seems to be fixed, so download like the docs say to.